Cambio 2 clase Manejador

+++ src/Patrones/Manejador.java
@@ -1,8 +1,8 @@
 package Patrones;
 
 public interface Manejador{
-    public void crearManejador(int monto, int denominacion);
+    public void crearManejador(double monto, double denominacion);
     public void setNext(Manejador m);
-    public boolean retirar(int monto);
-    public boolean depositar(int monto, int denominacion);
+    public boolean retirar(double monto);
+    public boolean depositar(double monto, double denominacion);
 }
